Music_store.sql
This project is about a music store data all over the world. This project involves querying a music store database to extract insightful information using SQL. The goal is to answer a series of questions. Each query is designed to retrieve specific information from the database, demonstrating SQL skills and an understanding of database schema relationships. This helps us to also understand which artist's albums were sold more, which customer invested the most on which artist and many more insights.